+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.9.0]
 - ILIAS 7 support
 - Remove generate deprecated PHP72Backport script
+- Remove generate deprecated GeneratePluginReadme
 - Remove generate ILIAS 5.4/PHP 7.0
 
 ## [2.8.6]